Roles & Permissions

+++ modules/RoleAndPermission/Resources/routes/api.php
@@ -2,26 +2,26 @@
 
 use Illuminate\Support\Facades\Route;
 use Modules\RoleAndPermission\Controllers\PermissionController;
-use Modules\RoleAndPermission\Controllers\RoleAndPermissionController;
 use Modules\RoleAndPermission\Controllers\RoleController;
+use Modules\RoleAndPermission\Enums\Permission;
 
-Route::group(['middleware' => ['auth:api'],"prefix"=>"roles"], function () {
-    Route::get('/', [RoleController::class, 'index']);
-    Route::post('/', [RoleController::class, 'store']);
-    Route::get('/{id}', [RoleController::class, 'show']);
-    Route::get('/{id}/permissions', [RoleController::class, 'getPermissions']);
-    Route::post('/{id}/assign-permissions', [RoleController::class, 'assignPermissionToRole']);
-    Route::put('/{id}', [RoleController::class, 'update']);
-    Route::delete('/{id}', [RoleController::class, 'delete']);
-});
+Route::group(['middleware' => ['auth:api', \Stancl\Tenancy\Middleware\InitializeTenancyByRequestData::class]], function () {
+    Route::get('permissions', [PermissionController::class, 'index'])->permission(Permission::PERMISSION_LIST());
+    Route::get('permissions/lookup', [PermissionController::class, 'permissionAsLookup']);
+    Route::post('permissions', [PermissionController::class, 'store'])->permission(Permission::PERMISSION_CREATE());
+    Route::get('permissions/{id}', [PermissionController::class, 'show'])->permission(Permission::PERMISSION_VIEW());
+    Route::put('permissions/{id}', [PermissionController::class, 'update'])->permission(Permission::PERMISSION_EDIT());
+    Route::delete('permissions/{id}', [PermissionController::class, 'delete'])->permission(Permission::PERMISSION_DELETE());
+    Route::patch('permissions/{id}/status', [PermissionController::class, 'setStatus'])->permission(Permission::PERMISSION_EDIT());
 
-
-Route::group(['middleware' => ['auth:api'],"prefix"=>"permissions"], function () {
-    Route::get('/', [PermissionController::class, 'index']);
-    Route::post('/', [PermissionController::class, 'store']);
-
-    Route::get('/{id}', [PermissionController::class, 'show']);
-    Route::put('/{id}', [PermissionController::class, 'update']);
-    Route::delete('/{id}', [PermissionController::class, 'delete']);
-
+    Route::group(['prefix' => 'roles'], function () {
+        Route::get('/', [RoleController::class, 'index'])->permission(Permission::ROLE_VIEW());
+        Route::post('/', [RoleController::class, 'store'])->permission(Permission::ROLE_CREATE());
+        Route::get('/{id}', [RoleController::class, 'show'])->permission(Permission::ROLE_VIEW());
+        Route::get('/{id}/permissions', [RoleController::class, 'getPermissions'])->permission(Permission::PERMISSION_VIEW());
+        Route::post('/{id}/assign-permissions', [RoleController::class, 'assignPermissionToRole']);
+        Route::put('/{id}', [RoleController::class, 'update'])->permission(Permission::ROLE_EDIT());
+        Route::delete('/{id}', [RoleController::class, 'delete'])->permission(Permission::ROLE_DELETE());
+        Route::patch('/{id}/status', [RoleController::class, 'setStatus'])->permission(Permission::ROLE_EDIT());
+    });
 });
